nixos-module: disable programs.command-not-found by default
Our module enables the `nix-index` module in nixpkgs by default. That module
asserts that `programs.command-not-found.enabled` must be false. So it would
make sense for our module to also disable `programs.command-not-found` by
default, since the base default value for it is true (as defined in the
`command-not-found` module's code itself).